uint64 for userid

+++ internal/model/user.go
@@ -1,6 +1,6 @@
 package model
 
 type User struct {
-	ID       uint   `gorm:"primaryKey;autoIncrement"`
+	ID       uint64 `gorm:"primaryKey;autoIncrement"`
 	Username string `gorm:"type:varchar(50);uniqueIndex;not null"`
 }
